Fix: Moved md5() to md5.js, pack the function, etc.
