+++ index.html
@@ -1,8 +1,6 @@
 <!DOCTYPE html>
 <head>
     <title>Demo Page</title>
-    <script src="md5.js"></script>
-    <script src="main.js"></script>
 </head>
 <body>
 
@@ -10,4 +8,5 @@
     <button onclick="test()">submit</button>
         <div id="ok"></div>
 <button onclick="getCookie()">Check Cookie</button>
+<script src="./main.js"></script>
 </body>
